split CA verification code into multiple files
